動画カードセクションをコンポーネントに切り出し
